backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
firebase-admin==6.6.0
reportlab==4.2.2

# Resume Screening Dependencies
pdfplumber==0.10.3
PyPDF2==3.0.1
spacy==3.7.2
language-tool-python==2.8.1
pandas==2.1.4
numpy==1.26.2
google-generativeai==0.8.5


